Move TemplateInformation models to separate file in models folder
